fix(svelte-check): warningFilter review follow-ups
- Truthiness parity: the sidecar now drops on any falsy predicate return
  (`Boolean(filter(w))`), matching Svelte's `if (!warning_filter(w)) return;`,
  not only a strict `false`.
- Ordering: apply warningFilter BEFORE `--compiler-warnings` promotion so a
  filtered-away warning can't be promoted to an error (official order is
  filter-then-promote) — was surfacing as an error and could flip the exit code.
- The one-time "filter unapplied" stderr notes are now `Once`-guarded so
  `--watch` doesn't re-print them every rebuild.
- Moved the Node-dependent `apply()` tests out of the `--lib` unit tests (whose
  CI job has no Node) into the integration suite; the lib keeps only the pure
  `extract_framed` logic tests.
- Added e2e coverage: falsy-return drop, and warningFilter-wins-over-error-promotion.
- Documented that an inline `vite.config.*` plugin `warningFilter` is intentionally
  unsupported.
